@@ -1,5 +1,7 @@
 package de.swa.clv.constraints;
 
+import de.swa.clv.ErrorCodeControl;
+import de.swa.clv.UseType;
 import de.swa.clv.ValidationRules;
 import de.swa.clv.groups.ConditionsGroup;
 import de.swa.clv.groups.ConditionsTopGroup;
@@ -15,24 +17,26 @@ public class ConditionsTest {
     ConditionsTopGroup someTopGroup = ConditionsTopGroup.AND(
             ConditionsGroup.OR(Condition.of("prop", Equals.none("ZOO")))
     );
+    ErrorCodeControl errorCodeControl = ErrorCodeControl.of(UseType.AS_SUFFIX, "#suffix");
     String someConstraintJson = "'constraint':{'type':'EQUALS_ANY','values':['FOO']}";
     String somePermissionsJson = "'permissions':{'type':'ALL','values':['BAR']}";
     String someTopGroupJson = "'condition':{'property':'prop','constraint':{'type':'EQUALS_NONE','values':['ZOO']}}";
+    String errorCodeControlJson = "'errorCodeControl':{'useType':'AS_SUFFIX','code':'#suffix'}";
 
     @Test
-    public void serializeToJson_constraintAndPermissionsAndConditions() {
-        Conditions conditions = new Conditions(someConstraint, somePermissions, someTopGroup);
+    public void serializeToJson_constraintAndPermissionsAndConditionsAndErrorCodeControl() {
+        Conditions conditions = new Conditions(someConstraint, somePermissions, someTopGroup, errorCodeControl);
 
         String json = Conditions.serializeToJson(conditions);
 
         String expected = Util.doubleQuote("{" + someConstraintJson + "," + somePermissionsJson + "," +
-                someTopGroupJson + "}");
+                someTopGroupJson + "," + errorCodeControlJson + "}");
         assertEquals(expected, json);
     }
 
     @Test
     public void serializeToJson_NoConstraint() {
-        Conditions conditions = new Conditions(ValidationRules.NO_CONSTRAINT, somePermissions, someTopGroup);
+        Conditions conditions = new Conditions(ValidationRules.NO_CONSTRAINT, somePermissions, someTopGroup, null);
 
         String json = Conditions.serializeToJson(conditions);
 
@@ -42,7 +46,7 @@ public class ConditionsTest {
 
     @Test
     public void serializeToJson_NoPermissions() {
-        Conditions conditions = new Conditions(someConstraint, ValidationRules.NO_PERMISSIONS, someTopGroup);
+        Conditions conditions = new Conditions(someConstraint, ValidationRules.NO_PERMISSIONS, someTopGroup, null);
 
         String json = Conditions.serializeToJson(conditions);
 
@@ -53,7 +57,7 @@ public class ConditionsTest {
     @Test
     public void serializeToJson_NoConditions() {
         Conditions conditions = new Conditions(someConstraint, somePermissions,
-                ValidationRules.NO_CONDITIONS_TOP_GROUP);
+                ValidationRules.NO_CONDITIONS_TOP_GROUP, null);
 
         String json = Conditions.serializeToJson(conditions);
 
@@ -64,7 +68,7 @@ public class ConditionsTest {
     @Test
     public void serializeToJson_constraintOnly() {
         Conditions conditions = new Conditions(someConstraint, ValidationRules.NO_PERMISSIONS,
-                ValidationRules.NO_CONDITIONS_TOP_GROUP);
+                ValidationRules.NO_CONDITIONS_TOP_GROUP, null);
 
         String json = Conditions.serializeToJson(conditions);
 
@@ -75,7 +79,7 @@ public class ConditionsTest {
     @Test
     public void serializeToJson_permissionsOnly() {
         Conditions conditions = new Conditions(ValidationRules.NO_CONSTRAINT, somePermissions,
-                ValidationRules.NO_CONDITIONS_TOP_GROUP);
+                ValidationRules.NO_CONDITIONS_TOP_GROUP, null);
 
         String json = Conditions.serializeToJson(conditions);
 
@@ -86,7 +90,7 @@ public class ConditionsTest {
     @Test
     public void serializeToJson_conditionsOnly() {
         Conditions conditions = new Conditions(ValidationRules.NO_CONSTRAINT, ValidationRules.NO_PERMISSIONS,
-                someTopGroup);
+                someTopGroup, null);
 
         String json = Conditions.serializeToJson(conditions);
 
